Moved code from keymap
